Added logic to validate the blocks based on the checkpoint and close connection on invalid blocks
Currently there is no checkpoint validation which might result
inconsistency of blocks. As Magic number is same with other forks of
bitcoin this might occur when a peer of other system sends the blocks
with longest chain. With this change we validate with existing
checkpoint height and hash, in case of mismatch we close the connection
and make it blacklisted.